+++ bin/index.js
@@ -1,21 +1,22 @@
 #!/usr/bin/env node
-'use strict';
 
-const {join} = require('path');
-const {cliBasics} = require('command-line-basics');
-const mainScript = require('../src/coveradge.js');
+import {join, dirname} from 'path';
+import {fileURLToPath} from 'url';
 
-const optionDefinitions = cliBasics(
+import {cliBasics} from 'command-line-basics';
+import mainScript from '../src/coveradge.cjs';
+
+const __dirname = dirname(fileURLToPath(import.meta.url));
+
+const optionDefinitions = await cliBasics(
   join(__dirname, '../src/optionDefinitions.js')
 );
 
 if (!optionDefinitions) { // cliBasics handled
-  process.exit();
+  process.exit(0);
 }
-(async () => {
 try {
   await mainScript(optionDefinitions);
 } catch (err) {
   console.error(err);
 }
-})();