+++ test-helpers/sample.cjs
@@ -1,0 +1,11 @@
+/* eslint-disable no-console -- Testing */
+'use strict';
+
+const a = 0;
+
+/* c8 ignore next 3 */
+if (a) {
+  console.log('A branch test');
+} else {
+  console.log('B branch test');
+}